pyproject.toml
[tool.poetry]
name = "django_nested_admin_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "<3"
django-polymorphic = "^2.1.2"
django-nested-admin = "^3.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
